client: primary p2p file transfer support
	- implements continuous file transfer
	  between peers.
	- this does not have any support for
	  choking/unchoking connections.

+++ src/p2p/MessageType.java
@@ -10,7 +10,8 @@ enum MessageType {
 	REQUEST,
 	PIECE;
 
-	public static MessageType getType(byte typeIndex) {
+	// getType does a reverse lookup on the MessageType.
+	public static MessageType getType(int typeIndex) {
 		return MessageType.values()[typeIndex];
 	}
 }
